Move high address mapping from config to memory
Of course it wont do to only map register addresses that appear in "memory.xml" (RegisterConfig). Because, for example, READREG/WRITEREG commands will use the address they find in the device description file ("camera.xml"). So the common point is RegisterMemory where data is get/set from memory. So move the mapping function there.
Also fixed the mapping! Was shifting one byte too far.
Also compensated for the device XML file writing, which has an absolute address that *shouldn't* be mapped.

Note I also looked into not doing the mapping in code, because it remains fragile. The proper fix would be to use a dictionary of register addresses, rather than a flat, contiguous memory space. From my reading of the GenICam standard so far (the one that defines the device description xml), an application will only ever read/write addresses that appear in there, up to the sizes defined in there. Alas, building up a list of addresses from the xml is a huge undertaking! Register addresses can be defined based on other variables which index other enums which reference other formulas which rely on other integer values.

+++ src/Server.cs
@@ -22,7 +22,7 @@
         private StreamClient streamClient = new StreamClient();
 
         /// <summary>
-        /// Callback that is triggere when ever a new Image need to be acquire
+        /// Callback that is triggered when ever a new Image need to be acquire
         /// </summary>
         private Func<ImageData>? onAcquiesceImageCallback;
 
@@ -121,12 +121,12 @@
             // an Internet URL or local file path, instead of relying on device memory. Note that the URLs
             // can also be specified in a manifest table instead, in which case these URLs are ignored.
             // The file can be large (eg. 600kB - ZIP support not implemented), so pick somewhere with clear air.
-            const int XML_FILE_ADDRESS = 0x01000000; //Spec says address must be aligned to 32-bit boundary.
+            const uint XML_FILE_ADDRESS = 0x01000000; //Spec says address must be aligned to 32-bit boundary.
             registers.WriteString(eBootstrapRegister.XML_Device_Description_File_First_URL,
-                "Local:camera.xml;" + ToHexString(XML_FILE_ADDRESS) + ";" + ToHexString(this.xml.Length));
+                "Local:camera.xml;" + ToHexString(XML_FILE_ADDRESS) + ";" + ToHexString((uint)this.xml.Length));
             // If the first fails, the second is used. But we have no other option! So just try the same.
             registers.WriteString(eBootstrapRegister.XML_Device_Description_File_Second_URL,
-                "Local:camera.xml;" + ToHexString(XML_FILE_ADDRESS) + ";" + ToHexString(this.xml.Length));
+                "Local:camera.xml;" + ToHexString(XML_FILE_ADDRESS) + ";" + ToHexString((uint)this.xml.Length));
             registers.WriteBytes(XML_FILE_ADDRESS, this.xml); //Store the file.
 
             //Finally, write memory.xml values over the top.
@@ -282,7 +282,7 @@
             return null;
         }
 
-        private string ToHexString(int num)
+        private string ToHexString(uint num)
         {
             return num.ToString("X");
         }
